server/requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-JWT-Extended>=4.6.0
psycopg2-binary>=2.9.10
Werkzeug>=3.0.0
python-dotenv==1.2.1